src/main/java/com/example/telegrambot/command/CommandContainer.java
package com.example.telegrambot.command;

import com.example.telegrambot.Services.SendBotMessageService;
import com.google.common.collect.ImmutableMap;

import java.util.Map;

import static com.example.telegrambot.command.CommandName.*;

public class CommandContainer {
    private final ImmutableMap <String,Command> commandMap;
    private final Command unknownCommand;

    public CommandContainer(SendBotMessageService sendBotMessageService) {
        commandMap = ImmutableMap.<String, Command>builder().
                put(START.getCommandName(),new StartCommand(sendBotMessageService))
                .put(STOP.getCommandName(),new StopCommand(sendBotMessageService))
                .put(HELP.getCommandName(),new HelpCommand(sendBotMessageService))
                .put(NO.getCommandName(),new NoCommand(sendBotMessageService)).build();
        this.unknownCommand = new UnknownCommand(sendBotMessageService);
    }

    public Command retrieveCommand(String commandIdentifier){
        return commandMap.getOrDefault(commandIdentifier,unknownCommand);
    }
}
